Moved Demo components and changed path references
